Refactor benchmarks and improve reporting
